fix: refactor & implemented requested changes
