+++ README.md
@@ -26,11 +26,11 @@ This repository houses plugins that Rollup considers critical to every day use o
 | [beep](packages/beep)                               | System beeps on errors and warnings                                                       |
 | [buble](packages/buble)                             | Compile ES2015 with buble                                                                 |
 | [commonjs](packages/commonjs)                       | Convert CommonJS modules to ES6                                                           |
-| [commonjs-shim](packages/commonjs-shim)             | Replace cjs syntax for esm output bundles                                                 |
 | [data-uri](packages/data-uri)                       | Import modules from Data URIs                                                             |
 | [dsv](packages/dsv)                                 | Convert .csv and .tsv files into JavaScript modules with d3-dsv                           |
 | [dynamic-import-vars](packages/dynamic-import-vars) | Resolving dynamic imports that contain variables.                                         |
 | [eslint](packages/eslint)                           | Verify entry point and all imported files with ESLint                                     |
+| [esm-shim](packages/esm-shim)                       | Replace cjs syntax for esm output bundles                                                 |
 | [graphql](packages/graphql)                         | Convert .gql/.graphql files to ES6 modules                                                |
 | [html](packages/html)                               | Create HTML files to serve Rollup bundles                                                 |
 | [image](packages/image)                             | Import JPG, PNG, GIF, SVG, and WebP files                                                 |
