--- packages/commonjs-shim/rollup.config.mjs
@@ -1,10 +1,0 @@
-import { readFileSync } from 'fs';
-
-import { createConfig } from '../../shared/rollup.config.mjs';
-
-export default {
-  ...createConfig({
-    pkg: JSON.parse(readFileSync(new URL('./package.json', import.meta.url), 'utf8'))
-  }),
-  input: 'src/index.ts'
-};
