+++ packages/esm-shim/rollup.config.mjs
@@ -1,0 +1,10 @@
+import { readFileSync } from 'node:fs';
+
+import { createConfig } from '../../shared/rollup.config.mjs';
+
+export default {
+  ...createConfig({
+    pkg: JSON.parse(readFileSync(new URL('./package.json', import.meta.url), { encoding: 'utf-8' }))
+  }),
+  input: 'src/index.ts'
+};
