+++ packages/esm-shim/src/index.ts
@@ -3,8 +3,8 @@
  * The core of this plugin was conceived by pi0 and is taken from the following repository:
  * https://github.com/unjs/unbuild/blob/main/src/builder/plugins/cjs.ts
  */
-import { commonjsShim } from './module';
+import { esmShim } from './module';
 
-export * from './type';
+export * from './types';
 export * from './utils';
-export default commonjsShim;
+export default esmShim;
