+++ packages/esm-shim/src/module.ts
@@ -1,0 +1,16 @@
+import type { Plugin } from 'rollup';
+
+import { provideCJSSyntax } from './utils';
+
+export function esmShim(): Plugin {
+  return {
+    name: 'esm-shim',
+    renderChunk(code, chunk, opts) {
+      if (opts.format === 'es') {
+        return provideCJSSyntax(code);
+      }
+
+      return null;
+    }
+  };
+}
